+++ tap-snapshots/lib-processRequest.test.mjs-TAP.test.js
@@ -206,7 +206,7 @@ exports[`lib/processRequest.test.mjs TAP Exceed max field size. Koa middleware. 
 exports[`lib/processRequest.test.mjs TAP Exceed max file size. Express middleware. Upload A. > Stream error. 1`] = `
 {
   "name": "PayloadTooLargeError",
-  "message": "File truncated as it exceeds the size limit.",
+  "message": "File truncated as it exceeds the 1 byte size limit.",
   "status": 413,
   "statusCode": 413,
   "expose": true
@@ -224,7 +224,7 @@ exports[`lib/processRequest.test.mjs TAP Exceed max file size. Express middlewar
 exports[`lib/processRequest.test.mjs TAP Exceed max file size. Koa middleware. Upload A. > Stream error. 1`] = `
 {
   "name": "PayloadTooLargeError",
-  "message": "File truncated as it exceeds the size limit.",
+  "message": "File truncated as it exceeds the 1 byte size limit.",
   "status": 413,
   "statusCode": 413,
   "expose": true